Update tests and move kubeflowspec
