+++ tests/core/test_component_specs.py
@@ -7,11 +7,7 @@ from unittest.mock import patch
 import pyarrow as pa
 import pytest
 import yaml
-from fondant.core.component_spec import (
-    ComponentSpec,
-    KubeflowComponentSpec,
-    OperationSpec,
-)
+from fondant.core.component_spec import ComponentSpec, OperationSpec
 from fondant.core.exceptions import InvalidComponentSpec
 from fondant.core.schema import Type
 
@@ -27,12 +23,6 @@ def valid_fondant_schema() -> dict:
 @pytest.fixture()
 def valid_fondant_schema_no_args() -> dict:
     with open(component_specs_path / "valid_component_no_args.yaml") as f:
-        return yaml.safe_load(f)
-
-
-@pytest.fixture()
-def valid_kubeflow_schema() -> dict:
-    with open(component_specs_path / "kubeflow_component.yaml") as f:
         return yaml.safe_load(f)
 
 
@@ -97,17 +87,6 @@ def test_attribute_access(valid_fondant_schema):
     )
 
 
-def test_kfp_component_creation(valid_fondant_schema, valid_kubeflow_schema):
-    """Test that the created kubeflow component matches the expected kubeflow component."""
-    fondant_component = ComponentSpec.from_dict(valid_fondant_schema)
-    kubeflow_component = KubeflowComponentSpec.from_fondant_component_spec(
-        fondant_component,
-        command=["fondant", "execute", "main"],
-        image_uri="example_component:latest",
-    )
-    assert kubeflow_component._specification == valid_kubeflow_schema
-
-
 def test_component_spec_no_args(valid_fondant_schema_no_args):
     """Test that a component spec without args is supported."""
     fondant_component = ComponentSpec.from_dict(valid_fondant_schema_no_args)
@@ -132,33 +111,11 @@ def test_component_spec_to_file(valid_fondant_schema):
         assert written_data == valid_fondant_schema
 
 
-def test_kubeflow_component_spec_to_file(valid_kubeflow_schema):
-    """Test that the KubeflowComponentSpec can be written to a file."""
-    kubeflow_component_spec = KubeflowComponentSpec(valid_kubeflow_schema)
-
-    with tempfile.TemporaryDirectory() as temp_dir:
-        file_path = os.path.join(temp_dir, "kubeflow_component_spec.yaml")
-        kubeflow_component_spec.to_file(file_path)
-
-        with open(file_path) as f:
-            written_data = yaml.safe_load(f)
-
-        # check if the written data is the same as the original data
-        assert written_data == valid_kubeflow_schema
-
-
 def test_component_spec_repr(valid_fondant_schema):
     """Test that the __repr__ method of ComponentSpec returns the expected string."""
     fondant_component = ComponentSpec.from_dict(valid_fondant_schema)
     expected_repr = f"ComponentSpec({valid_fondant_schema!r})"
     assert repr(fondant_component) == expected_repr
-
-
-def test_kubeflow_component_spec_repr(valid_kubeflow_schema):
-    """Test that the __repr__ method of KubeflowComponentSpec returns the expected string."""
-    kubeflow_component_spec = KubeflowComponentSpec(valid_kubeflow_schema)
-    expected_repr = f"KubeflowComponentSpec({valid_kubeflow_schema!r})"
-    assert repr(kubeflow_component_spec) == expected_repr
 
 
 def test_component_spec_generic_consumes(valid_fondant_schema_generic_consumes):
